requirements.txt
Django==1.9.9
psycopg2-binary==2.7.7
gunicorn
python-dotenv
celery==3.1.25
redis==2.10.6